+++ package.json
@@ -53,6 +53,7 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "lint": "eslint \"src/**/*.{ts,js,json}\"",
+    "typecheck": "tsc --noEmit",
     "format": "prettier --write \"src/**/*.ts\"",
     "lint:fix": "npm run lint -- --fix",
     "build:dev": "microbundle watch",
